+++ crosschain-contracts/contracts/CrossChainEnabledAMB.sol
@@ -19,7 +19,7 @@ abstract contract CrossChainEnabledAMB is CrossChainEnabled {
         return IAMB(amb).messageSender();
     }
 
-    function _crossChainCall(address target, bytes memory data, uint256 gas) internal virtual override returns (bool) {
+    function _crossChainCall(address target, bytes memory data, uint32 gas) internal virtual override returns (bool) {
         require(IAMB(amb).maxGasPerTx() <= gas);
         IAMB(amb).requireToPassMessage(target, data, gas);
         return true;
